backend/recommender/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mangum==0.17.0

# Database
supabase==2.3.4
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# ML (para futuras fases)
scikit-learn==1.4.0
scipy==1.11.4

# Utilidades
pydantic==2.5.3
httpx==0.26.0
